requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
cloudscraper==1.2.71
colorama==0.4.6
cryptography==46.0.7
distro==1.9.0
fastapi==0.136.0
google-auth==2.49.2
google-genai==1.73.1
greenlet==3.4.0
groq==1.2.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.13
lxml==6.1.0
playwright==1.58.0
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
pypdf==6.10.2
pyee==13.0.1
pyparsing==3.3.2
python-dotenv==1.2.2
PyYAML==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
sniffio==1.3.1
soupsieve==2.8.3
starlette==1.0.0
tenacity==9.1.4
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.45.0
watchfiles==1.1.1
websockets==16.0
